only glob on ewf files in loadiosubsys
Fixes a problem where mmls popup was broken on pyflag_stdimage_0.x because it
ended with a digit and that was the test for encase files....
How can this be done smarter?? In many cases, the behavior prior to this patch
may be desired, e.g. in split-dd files.

darcs-hash:20080129112758-84eb5-5d5eab847b21259c04716bd0a1090c5476f104f9.gz